Small refactoring and cleanup
