+++ services/headless-lms/server/src/programs/seed/seed_file_storage.rs
@@ -1,3 +1,4 @@
+use super::certificate_fonts_data::CERTIFICATE_FONTS;
 use headless_lms_utils::file_store::{FileStore, local_file_store::LocalFileStore};
 use std::path::Path;
 
@@ -5,109 +6,6 @@ const REPOSITORY_EXERCISE_1: &[u8] = include_bytes!("./data/repository-exercise-
 const REPOSITORY_EXERCISE_2: &[u8] = include_bytes!("./data/repository-exercise-2.tar.zst");
 const CERTIFICATE_BACKGROUND: &[u8] = include_bytes!("./data/certificate-background.svg");
 const AUTHOR_IMAGE: &[u8] = include_bytes!("./data/lilo-and-stitch.jpg");
-
-// Build fails on unhydrated git-lfs pointers, which include_bytes! would embed silently.
-const _: () = {
-    let mut i = 0;
-    while i < CERTIFICATE_FONTS.len() {
-        assert!(
-            matches!(
-                CERTIFICATE_FONTS[i].1,
-                [0x00, 0x01, 0x00, 0x00, ..] | [b'O', b'T', b'T', b'O', ..]
-            ),
-            "certificate font is not a valid font file (likely an unhydrated git-lfs pointer); run: git lfs pull --include=\"services/headless-lms/server/src/programs/seed/data\""
-        );
-        i += 1;
-    }
-};
-
-/// Certificate fonts, as `(file_path, bytes)`. Paths must match `seed_certificate_fonts`.
-const CERTIFICATE_FONTS: &[(&str, &[u8])] = &[
-    (
-        "fonts/inter-variable.ttf",
-        include_bytes!("./data/InterVariable.ttf"),
-    ),
-    (
-        "fonts/noto-sans-cjk-sc.otf",
-        include_bytes!("./data/NotoSansCJKsc-Regular.otf"),
-    ),
-    (
-        "fonts/noto-sans-arabic.ttf",
-        include_bytes!("./data/NotoSansArabic.ttf"),
-    ),
-    (
-        "fonts/noto-sans-hebrew.ttf",
-        include_bytes!("./data/NotoSansHebrew.ttf"),
-    ),
-    (
-        "fonts/noto-sans-thai.ttf",
-        include_bytes!("./data/NotoSansThai.ttf"),
-    ),
-    (
-        "fonts/noto-sans-devanagari.ttf",
-        include_bytes!("./data/NotoSansDevanagari.ttf"),
-    ),
-    (
-        "fonts/noto-sans-bengali.ttf",
-        include_bytes!("./data/NotoSansBengali.ttf"),
-    ),
-    (
-        "fonts/noto-sans-tamil.ttf",
-        include_bytes!("./data/NotoSansTamil.ttf"),
-    ),
-    (
-        "fonts/noto-sans-telugu.ttf",
-        include_bytes!("./data/NotoSansTelugu.ttf"),
-    ),
-    (
-        "fonts/noto-sans-kannada.ttf",
-        include_bytes!("./data/NotoSansKannada.ttf"),
-    ),
-    (
-        "fonts/noto-sans-malayalam.ttf",
-        include_bytes!("./data/NotoSansMalayalam.ttf"),
-    ),
-    (
-        "fonts/noto-sans-gujarati.ttf",
-        include_bytes!("./data/NotoSansGujarati.ttf"),
-    ),
-    (
-        "fonts/noto-sans-gurmukhi.ttf",
-        include_bytes!("./data/NotoSansGurmukhi.ttf"),
-    ),
-    (
-        "fonts/noto-sans-oriya.ttf",
-        include_bytes!("./data/NotoSansOriya.ttf"),
-    ),
-    (
-        "fonts/noto-sans-sinhala.ttf",
-        include_bytes!("./data/NotoSansSinhala.ttf"),
-    ),
-    (
-        "fonts/noto-sans-armenian.ttf",
-        include_bytes!("./data/NotoSansArmenian.ttf"),
-    ),
-    (
-        "fonts/noto-sans-georgian.ttf",
-        include_bytes!("./data/NotoSansGeorgian.ttf"),
-    ),
-    (
-        "fonts/noto-sans-ethiopic.ttf",
-        include_bytes!("./data/NotoSansEthiopic.ttf"),
-    ),
-    (
-        "fonts/noto-sans-khmer.ttf",
-        include_bytes!("./data/NotoSansKhmer.ttf"),
-    ),
-    (
-        "fonts/noto-sans-lao.ttf",
-        include_bytes!("./data/NotoSansLao.ttf"),
-    ),
-    (
-        "fonts/noto-sans-myanmar.ttf",
-        include_bytes!("./data/NotoSansMyanmar.ttf"),
-    ),
-];
 
 #[derive(Clone)]
 pub struct SeedFileStorageResult {}
@@ -135,7 +33,7 @@ pub async fn seed_file_storage() -> anyhow::Result<SeedFileStorageResult> {
             "application/octet-stream",
         )
         .await?;
-    for &(path, bytes) in CERTIFICATE_FONTS {
+    for &(_, path, bytes) in CERTIFICATE_FONTS {
         file_storage
             .upload(Path::new(path), bytes.to_vec(), "application/octet-stream")
             .await?;
